New multi-proc matching with index-based search complete
After hitting a few snags and testing, this is the finished
implementaion of the 3rd generation link matcher. A new command line
flag is added to control contextual word length (for completeness of
customizability). Help is refactored into a short and long version.
Default for link matching is changed to 0.8 (from 0.7). Various hints
put into the code to help the author fine-tune the flags if necessary
for various potential scenarios.